edge/migrate_schema_check.py
from supabase_client import supabase

def migrate():
    print("Attempting to add 'method' column to 'access_logs'...")
    try:
        # Note: Supabase Python client doesn't support raw SQL easily unless RPC is used.
        # But we can try to insert a record with 'method' to see if it specifically triggers a schema error.
        # or we can check if there's a way to run SQL.
        
        # Since we don't have a direct SQL execution tool, we will rely on the resilient code.
        # However, I can try to use the 'rpc' method if a 'run_sql' function exists.
        
        # Most Supabase setups have an internal 'run_sql' or similar if configured, 
        # but it's unlikely for security reasons.
        
        print("Schema migration skipped - relying on application-level resilience.")
        
    except Exception as e:
        print(f"Migration failed: {e}")

if __name__ == "__main__":
    migrate()
