edge/test_api_response.py
import requests
import json

url = "http://localhost:8001/api/biometrics/face/verify"
# We need an image, but the error happens after match.
# Since we want to see the JSON structure, we can try to find an image or check the code.
# I will instead look at the running process logs if possible, or use a script to check keys.

print("Checking Biometric API response structure...")
# I'll just check if the name is in the result of a mock-like call if I can, 
# but better to just look at the code one more time very carefully.
