edge/requirements.txt
fastapi
uvicorn
supabase
face-recognition
pillow
numpy
python-dotenv
httpx
python-multipart
bleak
